requirements.txt
argcomplete==1.9.3
boto3==1.9.35
botocore==1.12.35
certifi==2018.10.15
cfn-flip==1.0.3
chardet==3.0.4
Click==7.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
future==0.16.0
hjson==3.0.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
marathon==0.10.0
MarkupSafe==1.0
placebo==0.8.2
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==3.13
requests==2.20.0
requests-toolbelt==0.8.0
s3transfer==0.1.13
six==1.11.0
slackclient==1.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.24
websocket-client==0.54.0
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.0
